shared/python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shared"
version = "0.1.0"
description = "Shared Python libraries for Clinical Voice AI Agent"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "structlog>=23.2.0",
    "redis[hiredis]>=5.0.0",
    "motor>=3.3.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-exporter-otlp>=1.22.0",
    "pybreaker>=1.2.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]
